+++ internal/delegate/repo.go
@@ -167,3 +167,31 @@ func (r *Repo) FindDelegatorsByVotes(votes []Vote) ([]summaryByVote, error) {
 
 	return result, nil
 }
+
+func (r *Repo) GetByFilters(filters []Filter) ([]Summary, error) {
+	db := r.db.Model(&Summary{})
+	for _, f := range filters {
+		db = f.Apply(db)
+	}
+
+	var list []Summary
+	if err := db.Find(&list).Error; err != nil {
+		return nil, fmt.Errorf("db.Find: %w", err)
+	}
+
+	return list, nil
+}
+
+func (r *Repo) GetCnt(filters []Filter) (int64, error) {
+	db := r.db.Model(&Summary{}).Joins("INNER JOIN daos ON daos.id::text = delegates_summary.dao_id")
+	for _, f := range filters {
+		db = f.Apply(db)
+	}
+
+	var cnt int64
+	if err := db.Count(&cnt).Error; err != nil {
+		return cnt, fmt.Errorf("db.Count: %w", err)
+	}
+
+	return cnt, nil
+}