+++ internal/delegate/service.go
@@ -301,7 +301,11 @@ func (s *Service) handleDelegate(_ context.Context, hr History) error {
 			return nil
 		}
 
+		addresses := make([]string, 0, len(hr.Delegations.Details)+1)
+		addresses = append(addresses, hr.AddressFrom)
+
 		for _, info := range hr.Delegations.Details {
+			addresses = append(addresses, info.Address)
 			if err = s.repo.CreateSummary(Summary{
 				AddressFrom:        strings.ToLower(hr.AddressFrom),
 				AddressTo:          strings.ToLower(info.Address),
@@ -313,6 +317,10 @@ func (s *Service) handleDelegate(_ context.Context, hr History) error {
 				return fmt.Errorf("createSummary [%s/%s/%s]: %w", hr.AddressFrom, info.Address, daoID.String(), err)
 			}
 		}
+
+		go func(list []string) {
+			s.ensResolver.AddRequests(list)
+		}(addresses)
 
 		return nil
 	}); err != nil {
@@ -393,4 +401,62 @@ func (s *Service) handleVotesCreated(ctx context.Context, batch []Vote) error {
 	}
 
 	return nil
+}
+
+// getAllDelegations returns list of delegations grouped by dao
+func (s *Service) getAllDelegations(_ context.Context, address string) (map[string][]Summary, error) {
+	list, err := s.repo.GetByFilters([]Filter{DelegatorFilter{Address: address}})
+	if err != nil {
+		return nil, fmt.Errorf("repo.GetByFilters: %w", err)
+	}
+
+	result := make(map[string][]Summary, len(list))
+	for _, info := range list {
+		if _, ok := result[info.DaoID]; !ok {
+			result[info.DaoID] = make([]Summary, 0, len(list))
+		}
+
+		result[info.DaoID] = append(result[info.DaoID], info)
+	}
+
+	return result, nil
+}
+
+// getAllDelegators returns list of delegators grouped by dao
+func (s *Service) getAllDelegators(_ context.Context, address string) (map[string][]Summary, error) {
+	list, err := s.repo.GetByFilters([]Filter{DelegationFilter{Address: address}})
+	if err != nil {
+		return nil, fmt.Errorf("repo.GetByFilters: %w", err)
+	}
+
+	result := make(map[string][]Summary, len(list))
+	for _, info := range list {
+		if _, ok := result[info.DaoID]; !ok {
+			result[info.DaoID] = make([]Summary, 0, len(list))
+		}
+
+		result[info.DaoID] = append(result[info.DaoID], info)
+	}
+
+	return result, nil
+}
+
+// getAllDelegators returns count of delegators based on address
+func (s *Service) getDelegatorsCnt(_ context.Context, address string) (int32, error) {
+	cnt, err := s.repo.GetCnt([]Filter{DelegationFilter{Address: address}})
+	if err != nil {
+		return 0, fmt.Errorf("repo.GetByFilters: %w", err)
+	}
+
+	return int32(cnt), nil
+}
+
+// getAllDelegators returns count of delegations based on address
+func (s *Service) getDelegationsCnt(_ context.Context, address string) (int32, error) {
+	cnt, err := s.repo.GetCnt([]Filter{DelegatorFilter{Address: address}})
+	if err != nil {
+		return 0, fmt.Errorf("repo.GetByFilters: %w", err)
+	}
+
+	return int32(cnt), nil
 }